Updated docstrings and altered decoder file structure
